#1010 [REFACTOR] useReport와 useReportHandller 파일 통합 -> useReport 파일 하나로 관리
